mavros/src/plugins/command.cpp: command_ack_timeout_ms int -> double

+++ mavros/src/plugins/command.cpp
@@ -64,7 +64,7 @@ public:
 		PluginBase::initialize(uas_);
 
 		cmd_nh.param("use_comp_id_system_control", use_comp_id_system_control, false);
-		int command_ack_timeout_ms;
+		double command_ack_timeout_ms;
 		cmd_nh.param("command_ack_timeout", command_ack_timeout_ms, ACK_TIMEOUT_MS_DEFAULT);
 		command_ack_timeout_dt = ros::Duration(command_ack_timeout_ms / 1000);
 
@@ -103,7 +103,7 @@ private:
 	bool use_comp_id_system_control;
 
 	L_CommandTransaction ack_waiting_list;
-	static constexpr int ACK_TIMEOUT_MS_DEFAULT = 5000;
+	static constexpr double ACK_TIMEOUT_MS_DEFAULT = 5000.0;
 	ros::Duration command_ack_timeout_dt;
 
 	/* -*- message handlers -*- */
